implement authorization at voucher page actions

+++ resources/views/voucher/list.blade.php
@@ -6,7 +6,9 @@
 <div class="card shadow">
   <div class="card-header py-3 d-flex flex-row align-items-center justify-content-between">
     <h6 class="m-0 font-weight-bold text-primary">Vouchers List</h6>
-    <a href="/voucher/form" class="btn btn-primary btn-sm"><i class="fas fa-plus"></i></a>
+    @if (permission(auth()->user(),'voucher','add'))
+      <a href="/voucher/form" class="btn btn-primary btn-sm"><i class="fas fa-plus"></i></a>
+    @endif
   </div>
   <div class="card-body">
     <table class="table table-bordered table-hover table-striped">
@@ -36,8 +38,12 @@
               </td>
               <td>{!! ($row->status) ? '<span class="badge badge-success">Active</span>' : '<span class="badge badge-secondary">Claimed</span>' !!}</td>
               <td>
-                <a href="/voucher/form/{{ $row->id }}" class="btn btn-warning btn-sm btn-icon"><i class="fas fa-pen"></i></a>
-                <a href="/voucher/delete?id={{ $row->id }}" onclick="return confirm('Are you sure want to delete this voucher ?')" class="btn btn-danger btn-sm btn-icon"><i class="fas fa-trash"></i></a>
+                @if (permission(auth()->user(),'transaction','edit'))
+                  <a href="/voucher/form/{{ $row->id }}" class="btn btn-warning btn-sm btn-icon"><i class="fas fa-pen"></i></a>
+                @endif
+                @if (permission(auth()->user(),'transaction','delete'))
+                  <a href="/voucher/delete?id={{ $row->id }}" onclick="return confirm('Are you sure want to delete this voucher ?')" class="btn btn-danger btn-sm btn-icon"><i class="fas fa-trash"></i></a>
+                @endif
               </td>
             </tr>
           @endforeach
